# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BookMyEvent API",
    "description": "Ticketing system API for event organisers to create & manage events.\nUsers can purchase/reserve tickets for events.\n",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@bookmyevent.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/events": {
      "post": {
        "operationId": "EventsController_create",
        "summary": "Create a new event",
        "tags": ["Events"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEventDto" },
              "example": {
                "name": "Summer Concert 2026",
                "description": "The biggest summer concert",
                "startTime": "2026-06-15T18:00:00.000Z",
                "endTime": "2026-06-15T23:00:00.000Z",
                "venueId": "550e8400-e29b-41d4-a716-446655440000"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "660e8400-e29b-41d4-a716-446655440000",
                    "name": "Summer Concert 2026",
                    "description": "The biggest summer concert",
                    "startTime": "2026-06-15T18:00:00.000Z",
                    "endTime": "2026-06-15T23:00:00.000Z",
                    "venueId": "550e8400-e29b-41d4-a716-446655440000",
                    "status": "DRAFT",
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  },
                  "message": "Event created successfully"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "EventsController_findAll",
        "summary": "Get all events",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "createdAt" }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "660e8400-e29b-41d4-a716-446655440000",
                      "name": "Summer Concert 2026",
                      "description": "The biggest summer concert",
                      "startTime": "2026-06-15T18:00:00.000Z",
                      "endTime": "2026-06-15T23:00:00.000Z",
                      "venueId": "550e8400-e29b-41d4-a716-446655440000",
                      "status": "DRAFT",
                      "createdAt": "2026-04-18T12:00:00.000Z"
                    }
                  ],
                  "pagination": { "skip": 0, "take": 10, "total": 1 }
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "summary": "Get an event by ID",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Event found",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "660e8400-e29b-41d4-a716-446655440000",
                    "name": "Summer Concert 2026",
                    "description": "The biggest summer concert",
                    "startTime": "2026-06-15T18:00:00.000Z",
                    "endTime": "2026-06-15T23:00:00.000Z",
                    "venueId": "550e8400-e29b-41d4-a716-446655440000",
                    "status": "DRAFT",
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": { "description": "Event not found" }
        }
      },
      "patch": {
        "operationId": "EventsController_update",
        "summary": "Update an event",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateEventDto" },
              "example": { "status": "PUBLISHED" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "660e8400-e29b-41d4-a716-446655440000",
                    "name": "Summer Concert 2026",
                    "status": "PUBLISHED"
                  },
                  "message": "Event updated successfully"
                }
              }
            }
          },
          "404": { "description": "Event not found" }
        }
      },
      "delete": {
        "operationId": "EventsController_remove",
        "summary": "Delete an event",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "204": { "description": "Event deleted successfully" },
          "404": { "description": "Event not found" }
        }
      }
    },
    "/events/{id}/seats/status": {
      "get": {
        "operationId": "EventsController_getSeatsStatus",
        "summary": "Get seat status for an event",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Seat status",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "totalSeats": 150,
                    "available": 120,
                    "reserved": 20,
                    "booked": 10
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "summary": "Create a new order",
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrderDto" },
              "example": {
                "userEmail": "user@example.com",
                "seatIds": [
                  "990e8400-e29b-41d4-a716-446655440000",
                  "990e8400-e29b-41d4-a716-446655440001"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "880e8400-e29b-41d4-a716-446655440000",
                    "userEmail": "user@example.com",
                    "totalAmount": 300,
                    "status": "RESERVED",
                    "expiresAt": "2026-04-18T12:30:00.000Z",
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  },
                  "message": "Order reserved. Please confirm within 10 minutes."
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "summary": "Get all orders",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "createdAt" }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "880e8400-e29b-41d4-a716-446655440000",
                      "userEmail": "user@example.com",
                      "totalAmount": 300,
                      "status": "RESERVED",
                      "expiresAt": "2026-04-18T12:30:00.000Z",
                      "createdAt": "2026-04-18T12:00:00.000Z"
                    }
                  ],
                  "pagination": { "skip": 0, "take": 10, "total": 1 }
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "summary": "Get an order by ID",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "880e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "880e8400-e29b-41d4-a716-446655440000",
                    "userEmail": "user@example.com",
                    "totalAmount": 300,
                    "status": "RESERVED",
                    "expiresAt": "2026-04-18T12:30:00.000Z",
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": { "description": "Order not found" }
        }
      },
      "patch": {
        "operationId": "OrdersController_update",
        "summary": "Update an order",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "880e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateOrderDto" },
              "example": { "status": "CONFIRMED" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "880e8400-e29b-41d4-a716-446655440000",
                    "status": "CONFIRMED"
                  },
                  "message": "Order updated successfully"
                }
              }
            }
          },
          "404": { "description": "Order not found" }
        }
      },
      "delete": {
        "operationId": "OrdersController_remove",
        "summary": "Delete an order",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "880e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "204": { "description": "Order deleted successfully" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/orders/{id}/confirm": {
      "post": {
        "operationId": "OrdersController_confirm",
        "summary": "Confirm an order",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "880e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Order confirmed successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "880e8400-e29b-41d4-a716-446655440000",
                    "status": "CONFIRMED"
                  },
                  "message": "Order confirmed successfully"
                }
              }
            }
          },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancel",
        "summary": "Cancel an order",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "880e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "880e8400-e29b-41d4-a716-446655440000",
                    "status": "CANCELLED"
                  },
                  "message": "Order cancelled successfully"
                }
              }
            }
          },
          "404": { "description": "Order not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Event name" },
          "description": {
            "type": "string",
            "description": "Event description"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "Event start time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "description": "Event end time"
          },
          "venueId": {
            "type": "string",
            "format": "uuid",
            "description": "Venue ID"
          },
          "status": { "type": "string", "description": "Event status" }
        },
        "required": ["name", "startTime", "endTime", "venueId"]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "endTime": { "type": "string", "format": "date-time" },
          "venueId": { "type": "string", "format": "uuid" },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PUBLISHED", "CANCELLED"]
          }
        }
      },
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "userEmail": {
            "type": "string",
            "format": "email",
            "description": "User email"
          },
          "seatIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "description": "Array of event seat IDs to reserve"
          }
        },
        "required": ["userEmail", "seatIds"]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["RESERVED", "CONFIRMED", "CANCELLED"]
          }
        }
      },
...
    }
  }
}